Provide a way to install a prototype object of private scripts
In order to make HTMLMarqueeElement in Blink-in-JS workable, we need a way to install a prototype object of the marquee element when the marquee element is created (e.g., when HTMLElementFactory creates the marquee element).

This CL implements PrivateScriptRunner::installClass, which is a C++ API to install a prototype object of a given class.

The API is tested in PrivateScriptTest.

BUG=341031

Review URL: https://codereview.chromium.org/373763002

git-svn-id: svn://svn.chromium.org/blink/trunk@177642 bbb929c8-8fbe-4397-9dbb-9b2b20218538