Oilpan: Do not repeatedly memset the same free list memory with zeroes.
The freelists are rebuild during sweeping, so all memory that is on
a freelist when we start sweeping will be needlessly memset.

Instead zero out the memory during sweeping when we know that memory
has become free.

This brings the time spent in memset from ~10% to ~3% on
EventsDispatchingInDeeplyNestedShadowTrees.

R=haraken@chromium.org, oilpan-reviews@chromium.org, wibling@chromium.org, zerny@chromium.org

Review URL: https://codereview.chromium.org/381053002

git-svn-id: svn://svn.chromium.org/blink/trunk@177844 bbb929c8-8fbe-4397-9dbb-9b2b20218538