Private scripts should support DOM exceptions
This CL is a preparation for making HTMLMarqueeElement workable in private scripts. This CL supports DOM exceptions thrown from private scripts.

It's a shame that we have to duplicate the list of ExceptionCodes in PrivateScriptRunner.js. Once we support internal APIs that allow private scripts (PrivateScriptRunner.js) to access C++ code (ExceptionCode.h), we might be able to remove the duplication.

TEST=private_script_unittest.html
BUG=341031

Review URL: https://codereview.chromium.org/384503007

git-svn-id: svn://svn.chromium.org/blink/trunk@177942 bbb929c8-8fbe-4397-9dbb-9b2b20218538