Rename WebSocket to DOMWebSocket
We have a naming conflict between blink::WebSocket and WebCore::WebSocket. Both
classes #define WebSocket_h, which would cause a problem if a file tried to
include both headers. This naming conflict will also cause a link error if we
end up merging the blink and WebCore namespaces.

This CL renames WebSocket to DOMWebSocket to clarify that this class is the
implementation of the DOM WebSocket interface. We use this convention in other
places to distinguish DOM objects from underlying platform concepts with the
same name (e.g., DOMWindow, DOMURL).

R=eseidel@chromium.org,tyoshino@chromium.org,yhirano@chromium.org

Review URL: https://codereview.chromium.org/394063002

git-svn-id: svn://svn.chromium.org/blink/trunk@178217 bbb929c8-8fbe-4397-9dbb-9b2b20218538