Improve the declared IDL type for postMessage()'s transferables
Make the postMessage() IDL type for the transferables argument be
equal to what the specs and the custom binding code actually
supports. That is, a sequence of Transferables.

Accurately defining the Transferable typedef is dependent on better
union type support (http://crbug.com/240176); provide an incomplete 
typedef until that is in place.

R=haraken@chromium.org,jsbell@chromium.org
BUG=393245

Review URL: https://codereview.chromium.org/393353003

git-svn-id: svn://svn.chromium.org/blink/trunk@178322 bbb929c8-8fbe-4397-9dbb-9b2b20218538