Remove Clang-on-Win fix for definitions of Color's static data members
Clang does treat the in-class initialization as a definition, just like MSVC.
The problem was that Clang failed to emit the definition unless it was referenced,
which broke the components build. This has been fixed in Clang r213304.

The work-around was added in Blink r176469, and this patch effectively removes that.

BUG=82385

Review URL: https://codereview.chromium.org/397323002

git-svn-id: svn://svn.chromium.org/blink/trunk@178401 bbb929c8-8fbe-4397-9dbb-9b2b20218538