Oilpan: WebAudio: Apply the weak HashMap pattern to remove an entry from AudioContext::m_dirtyAudioSummingJunctions.
r178238, which made AudioSummingJunction::m_context a weak member, didn't
work. If the AudioContext and an AudioSummingJunction object become unreachable
togeter, AudioSummingJunction::m_context is not cleared and their destuction
order is not deterministic. We must not touch weak members in destructors.

With this CL, we register all of AudioSummingJunction objects to AudioContext,
and applies the weak HashMap pattern to clear m_dirtyAudioSummingJunctions.

BUG=392788

Review URL: https://codereview.chromium.org/393133003

git-svn-id: svn://svn.chromium.org/blink/trunk@178521 bbb929c8-8fbe-4397-9dbb-9b2b20218538