Fix exceptions when running the tests.
-Make NetworkSimulator error when runTest calls are nested.
-Harden NetworkSimulator against exceptions.
-Fix incorrect nested runTest call in treestatus_unittests.js.
-Fix network requests in ct-results-by-builder-tests.html.

This exposes that some of the tests were not overriding net
and thus the tests now do a ton of network requests.
This will go away when https://codereview.chromium.org/402603007
lands, so it doesn't seem worth the effort to fix all the
broken cases.

NOTRY=true
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/401163002

git-svn-id: svn://svn.chromium.org/blink/trunk@178529 bbb929c8-8fbe-4397-9dbb-9b2b20218538