Fix a bunch of UI nits in sheriff-o-matic.
-Add back the dividers between failure cards.
The wall of text and buttons was overwhelming.
-Fix the heights of things so that the page contents
isn't all overflowing out of height:0 containers.
-Remove the useless divider in ct-builder-grid.
-Make the buttons look like esprehn's codereview tool.
-Add a new sheriff-friendly favicon.
-Have the results panel consist of two overflow:auto
regions. One for the menu and one for the results.
-Get rid of tabs on top. Instead have a header that
converts to a "go back" button when you are in the
results page.
-Have ct-tree-status be collapsed to 0 height when it's empty.
To make this work, have the default status be the empty
string so that the status attribute isn't removed when the value
is open.
-Fix ellipsizing CSS in ct-tree-status

NOTRY=true
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/400353006

git-svn-id: svn://svn.chromium.org/blink/trunk@178539 bbb929c8-8fbe-4397-9dbb-9b2b20218538