Move commit list and builder latest revisions up to ct-unexpected-failures
Create a model directory for model classes. We should never be using
dumb objects as our model (even in tests!). Having explicit classes makes
more clear the types of the objects a component takes in. Also,
it gives a clear location for putting logic related to that bit of the
model.

Moving the commit list and the builder revisions up is a step in the
direction of enabling https://codereview.chromium.org/402603007. 

NOTRY=true

Review URL: https://codereview.chromium.org/404463003

git-svn-id: svn://svn.chromium.org/blink/trunk@178582 bbb929c8-8fbe-4397-9dbb-9b2b20218538