Fix flakiness dashboard loading for builds with git hashes
This patch fixes flakiness dashboard loading for builds with git hashes.
These hashes are not supported but we shouldn't completely break the
dashboard if they occur. This patch causes us to throw a nice error
message to the console and continue if we do hit a git hash.

The htmlForSingleTestRow test has also been updated to add the class
"file-new-bug". This tested regressed in
https://src.chromium.org/viewvc/blink?view=rev&revision=178396

BUG=395484
NOTRY=true

Review URL: https://codereview.chromium.org/405223003

git-svn-id: svn://svn.chromium.org/blink/trunk@178588 bbb929c8-8fbe-4397-9dbb-9b2b20218538