Enforce clearing renderers' paint invalidation state
The ASSERT checks that no flags cleared by clearPaintInvalidationState()
is still set up after calling repaint-after-compositing-update.

The ASSERT is skipped for SVG subtree as we have an optimization that
skips them. We could reproduce the exact condition but it seems not
very useful to do so.

The change caught several bugs (yay!):
- one in shouldCheckForPaintInvalidationAfterLayout()
that made the ASSERT trip on fast/repaint/table-row.html.
- one where setShouldDoFullPaintInvalidationAfterLayout didn't
mark the containing block chain for invalidation.

BUG=385169
R=dsinclair@chromium.org, leviw@chromium.org

Review URL: https://codereview.chromium.org/389573008

git-svn-id: svn://svn.chromium.org/blink/trunk@178719 bbb929c8-8fbe-4397-9dbb-9b2b20218538