flex shorthand incorrectly accepting 1 auto 1 (i.e. flex-basis in the middle)
This patch is in continuation with the given CL
https://codereview.chromium.org/396263003/
which fixes the bug of flex shorthand incorrectly accepting
values in which flex-basis is in the middle and does not have value
"auto". It does not took care of the case 1 auto 1 which according to
the grammar of te flex shorthand is incorrect.

This patch is adding support for the given missed case "1 auto 1"
by modifying the condition to check for if flex-basis is in the
middle or not in the argument list for all the different types of
values of flex-basis.

BUG=380984

Review URL: https://codereview.chromium.org/408683002

git-svn-id: svn://svn.chromium.org/blink/trunk@178738 bbb929c8-8fbe-4397-9dbb-9b2b20218538